requirements.txt
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Faker==36.1.1
fonttools==4.56.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
roman==5.0
scipy==1.15.2
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tqdm==4.67.1
tzdata==2025.1
